+++ src/GameLevel.cpp
@@ -157,7 +157,7 @@ void Game::newTick()
 void Game::setCurrentPower(Power* power)
 {
     powerHandle = power;
-    this->powerTime = powerHandle->maxPowerTime;
+    this->powerTime = powerHandle->maxPowerTime * this->abilities.calculateValue(PlayerAbilityManager::POWER_TIME);
     this->powerMaxTime = this->powerTime;
     this->powerCooldown = -1;
 
@@ -172,7 +172,7 @@ void Game::setCurrentPower(Power* power)
 }
 void Game::stopCurrentPower()
 {
-    this->powerCooldown = 1800; // 30 seconds
+    this->powerCooldown = 1800 / this->abilities.calculateValue(PlayerAbilityManager::POWER_COOLDOWN_TIME); // 30 seconds
     this->powerTime = -1; //0 - can use power, -1 - cooldown, >0 - power is used, 1 - set cooldown!
 
     // power 'stop'