+++ src/PowerFreeze.cpp
@@ -13,7 +13,7 @@ void PowerFreeze::onPowerStop()
 
 void PowerFreeze::onPowerTick(int tc)
 {
-    if(tc > maxPowerTime * 17 / 18)
+    if(Game::instance->powerMaxTime - tc < maxPowerTime / 18)
         Game::instance->setGameSpeed(Game::instance->getGameSpeed() / (0.025f * sqrt(currentLevel) + 1.f));
     else if(tc < maxPowerTime / 9)
         Game::instance->setGameSpeed(Game::instance->getGameSpeed() * (0.000725f * sqrt(currentLevel) + 1.f));