+++ CHANGELOG.md
@@ -4,15 +4,22 @@
 
 # BETA 0.2 SNAPSHOTS
 
++ Added Shop GUI with powers and abilities.
++ Added abilities: Power Cooldown Time, Power Time, Damage
+* Power indicator won't be displayed if any power wasn't equipped.
+* Made GUI font a bit smaller - 30 pt -> 28 pt
+* Made Freeze a bit shorter - 60 s -> 50 s
+* Changed Powers' GUI apperrance.
+* Point Boost time is shorter - 16.7 s -> 5 s
 ## v190806
 + You can have now only 2 powers at once.
 * Made profile format compatible with other profile format versions. Your powers won't be longer removed!
 * Fixed Map Selection bug.
-* Coins are earned by full destruction of car (not only damage).
+* Coins are earned by full destruction of car (not damage only).
 * Cars destroyed by power will give coins.
 * Armored Car won't be longer spawned double.
 * Oil is better on first level but its force rises slower.
-* Changed apperrance of Oil power.
+* Changed apperrance of Oil power (on higher levels).
 ## v190805
 * Changed background color of some maps
 * Map Selection now is set on last unlocked map (not always on Coutryside)
@@ -70,7 +77,7 @@
 * Made maps 4x more expensive due to point multiplier.
 * Now game speed depends on display.
 * Now car panes are always blue.
-* Now the cars are spawning in other timespan in different maps.
+* Now cars are spawning in other timespan in different maps.
 * Now the Oil power is working on 2 lanes (not 3).
 * Changed toggle button graphics (bigger slider).
 * Changed graphics (eg. power time indicator).